playbooks/meetings-exporter/src/requirements.txt
# Meeting Data Exporter Playbook — dependencies
# Install: pip install -r requirements.txt

requests>=2.28
google-api-python-client>=2.100
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.1.1
python-dotenv>=1.0
